+++ src/common.ts
@@ -38,7 +38,12 @@ export const createAtoms = <
     const observerCache = get(observerCacheAtom)
     let observer = observerCache.get(queryClient)
     if (observer) {
-      observer.setOptions(options)
+      // Needs to delay because this is called in render
+      // and `setOptions` notifies listeners.
+      // FIXME Is there a better way?
+      Promise.resolve().then(() => {
+        ;(observer as Observer).setOptions(options)
+      })
     } else {
       observer = createObserver(queryClient, options)
       observerCache.set(queryClient, observer)
@@ -83,7 +88,8 @@ export const createAtoms = <
     }
   )
 
-  const baseDataAtom = atomWithObservable((get) => {
+  const baseDataAtom = atom((get) => {
+    getOptions(get) // re-create observable when options change
     const observer = get(observerAtom)
     const observable = {
       subscribe: (
@@ -102,16 +108,18 @@ export const createAtoms = <
         return { unsubscribe }
       },
     }
-    return observable
+    const resultAtom = atomWithObservable(() => observable)
+    return resultAtom
   })
 
   const dataAtom = atom(
     (get) => {
-      const baseData = get(baseDataAtom)
-      if (baseData.error) {
-        throw baseData.error
+      const resultAtom = get(baseDataAtom)
+      const result = get(resultAtom)
+      if (result.error) {
+        throw result.error
       }
-      return baseData.data
+      return result.data
     },
     (_get, set, action: Action) => set(statusAtom, action)
   )